PowerStationStateConverter ready.

+++ src/main/java/com/epsm/electricPowerSystemDispatcher/repository/SavedGeneratorStateDao.java
@@ -5,6 +5,6 @@ import java.util.List;
 import com.epsm.electricPowerSystemDispatcher.model.domain.SavedGeneratorState;
 
 public interface SavedGeneratorStateDao {
-	public List<SavedGeneratorState> getStatesByPowerStationNumber(int powerStationNumber);
+	public List<SavedGeneratorState> getStatesByPowerStationNumber(long powerStationNumber);
 	public void saveState(SavedGeneratorState state);
 }
